+++ library/src/main/java/com/pengrad/telegrambot/model/suggestedposts/SuggestedPostInfo.kt
@@ -4,6 +4,6 @@ package com.pengrad.telegrambot.model.suggestedposts
 data class SuggestedPostInfo (
     @get:JvmName("state") val state: String,
     @get:JvmName("price") val price: SuggestedPostPrice?,
-    @get:JvmName("sendDate") val sendDate: Int?
+    @get:JvmName("sendDate") val sendDate: Long?
 ){
 }
